pyproject.toml
[project]
name = "marimo-pymc-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arviz-plots>=0.6.0",
    "marimo>=0.14.17",
    "pymc>=5.25.1",
    "ruff>=0.12.8",
]
